+++ src/modules/disk/option.h
@@ -4,15 +4,22 @@
 
 #include "common/option.h"
 
-typedef enum FFDiskType
+typedef enum FFDiskVolumeType
 {
-    FF_DISK_TYPE_NONE = 0,
-    FF_DISK_TYPE_REGULAR_BIT = 1 << 0,
-    FF_DISK_TYPE_HIDDEN_BIT = 1 << 1,
-    FF_DISK_TYPE_EXTERNAL_BIT = 1 << 2,
-    FF_DISK_TYPE_SUBVOLUME_BIT = 1 << 3,
-    FF_DISK_TYPE_UNKNOWN_BIT = 1 << 4,
-} FFDiskType;
+    FF_DISK_VOLUME_TYPE_NONE = 0,
+    FF_DISK_VOLUME_TYPE_REGULAR_BIT = 1 << 0,
+    FF_DISK_VOLUME_TYPE_HIDDEN_BIT = 1 << 1,
+    FF_DISK_VOLUME_TYPE_EXTERNAL_BIT = 1 << 2,
+    FF_DISK_VOLUME_TYPE_SUBVOLUME_BIT = 1 << 3,
+    FF_DISK_VOLUME_TYPE_UNKNOWN_BIT = 1 << 4,
+    FF_DISK_VOLUME_TYPE_READONLY_BIT = 1 << 5,
+} FFDiskVolumeType;
+
+typedef enum FFDiskCalcType
+{
+    FF_DISK_CALC_TYPE_FREE,
+    FF_DISK_CALC_TYPE_AVAILABLE,
+} FFDiskCalcType;
 
 typedef struct FFDiskOptions
 {
@@ -20,5 +27,6 @@ typedef struct FFDiskOptions
     FFModuleArgs moduleArgs;
 
     FFstrbuf folders;
-    FFDiskType showTypes;
+    FFDiskVolumeType showTypes;
+    FFDiskCalcType calcType;
 } FFDiskOptions;