+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 2.0.5
+
+Bugfixes:
+* Fix segfault when using libxrandr (#544, Display, Linux)
+* Don't print 0px (#544, Cursor)
+
+Features:
+* Add option `--disk-use-available` (#543)
+* Add option `--disk-show-readonly`
+
 # 2.0.4
 
 Bugfixes: